chore(lib/patch): moves browser-specific patches into lib/patch/browser
