+++ lib/patch/browser.js
@@ -2,14 +2,14 @@
 
 var fnPatch = require('./functions');
 var promisePatch = require('./promise');
-var mutationObserverPatch = require('./mutation-observer');
 var definePropertyPatch = require('./define-property');
-var registerElementPatch = require('./register-element');
-var webSocketPatch = require('./websocket');
-var eventTargetPatch = require('./event-target');
 var propertyDescriptorPatch = require('./property-descriptor');
-var geolocationPatch = require('./geolocation');
-var fileReaderPatch = require('./file-reader');
+var mutationObserverPatch = require('./browser/mutation-observer');
+var registerElementPatch = require('./browser/register-element');
+var webSocketPatch = require('./browser/websocket');
+var eventTargetPatch = require('./browser/event-target');
+var geolocationPatch = require('./browser/geolocation');
+var fileReaderPatch = require('./browser/file-reader');
 
 function apply() {
   fnPatch.patchSetClearFunction(global, [
