+++ lib/patch/browser/event-target.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var utils = require('../utils');
+var utils = require('../../utils');
 
 function apply() {
   // patched properties depend on addEventListener, so this needs to come first
