+++ lib/patch/browser/file-reader.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var utils = require('../utils');
+var utils = require('../../utils');
 
 function apply() {
   utils.patchClass('FileReader');
