+++ lib/patch/browser/geolocation.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var utils = require('../utils');
+var utils = require('../../utils');
 
 function apply() {
   if (global.navigator && global.navigator.geolocation) {
