+++ lib/patch/browser/mutation-observer.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var keys = require('../keys');
+var keys = require('../../keys');
 
 var originalInstanceKey = keys.create('originalInstance');
 var creationZoneKey = keys.create('creationZone');
