+++ lib/patch/browser/register-element.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var _redefineProperty = require('./define-property')._redefineProperty;
-var utils = require("../utils");
+var _redefineProperty = require('../define-property')._redefineProperty;
+var utils = require("../../utils");
 
 function apply() {
   if (utils.isWebWorker() || !('registerElement' in global.document)) {
