+++ lib/patch/browser/websocket.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var utils = require('../utils');
+var utils = require('../../utils');
 
 // we have to patch the instance since the proto is non-configurable
 function apply() {
