+++ lib/patch/property-descriptor.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var webSocketPatch = require('./websocket');
+var webSocketPatch = require('./browser/websocket');
 var utils = require('../utils');
 var keys = require('../keys');
